requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
drf-yasg==1.21.10